+++ src/actions/actions.js
@@ -5,9 +5,9 @@ import {
   fetchRestaurant,
   postLogin,
   postReview,
-} from './services/api';
+} from '../services/api';
 
-import { saveItem } from './services/storage';
+import { saveItem } from '../services/storage';
 
 export function setRegions(regions) {
   return {
