+++ src/container/RestaurantContainer.jsx
@@ -1,18 +1,19 @@
+/* eslint-disable react/jsx-no-bind */
 import { useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import RestaurantDetail from './RestaurantDetail';
-import ReviewForm from './ReviewForm';
-import Reviews from './Reviews';
+import RestaurantDetail from '../components/RestaurantDetail';
+import ReviewForm from '../components/ReviewForm';
+import Reviews from '../components/Reviews';
 
 import {
   loadRestaurant,
   changeReviewField,
   sendReview,
-} from './actions';
+} from '../actions/actions';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function RestaurantContainer({ restaurantId }) {
   const dispatch = useDispatch();
