+++ src/pages/LoginPage.jsx
@@ -1,4 +1,4 @@
-import LoginFormContainer from './LoginFormContainer';
+import LoginFormContainer from '../container/LoginFormContainer';
 
 export default function LoginPage() {
   return (
