+++ src/reducer/reducer.test.js
@@ -13,7 +13,7 @@ import {
   changeReviewField,
   clearReviewFields,
   setReviews,
-} from './actions';
+} from '../actions/actions';
 
 describe('reducer', () => {
   context('when previous state is undefined', () => {
