+++ src/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 
 import thunk from 'redux-thunk';
 
-import reducer from './reducer';
+import reducer from './reducer/reducer';
 
 const store = createStore(reducer, applyMiddleware(thunk));
